requirements.txt
absl-py==2.1.0
altair==5.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
azure-ai-vision-imageanalysis==1.0.0b2
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
easyocr==1.7.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
executing==2.0.1
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h5py==3.10.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
Keras-Applications==1.0.8
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msrest==0.7.1
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydeck==0.8.1b0
Pygments==2.17.2
pyodbc==5.1.0
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.2.12
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
validators==0.24.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
